requirements.txt
redis==5.2.1
fastapi==0.115.12
gunicorn==23.0.0
uvicorn==0.34.2
pydantic-settings==2.9.1
python-telegram-bot==22.0
python-dotenv==1.1.0
pytest==8.3.5
google-search-results==2.4.2
trafilatura==2.0.0
openai==1.76.0

# dev dependencies
black
ruff
isort
